package/go: ensure there is a host-go provider, take 2
Commit 7b2a164b7460 (package/go/go-bin: new host-go provider) broke the
case for prompt-less, host-only go packages, as the main host-go symbol
would not be selected.

This very same use-case was initially broken when the first go-src
provider was introduced with fa2536ec9401 (package/go: make host package
a virtual package), and subsequently fixed in 99a5d51c6e9f (package/go:
ensure there is a host-go provider), so 7b2a164b7460 is causing a
regression of a fixed regression.

Fix that the very same way it was originally fixed.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>
Cc: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Thomas Perale <thomas.perale@mind.be>
Signed-off-by: Julien Olivain <ju.o@free.fr>